requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alpaca-py==0.41.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
curl_cffi==0.12.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
eventkit==1.0.3
exchange_calendars==4.10.1
executing==2.0.1
fastjsonschema==2.21.1
Flask==3.1.2
fluids==1.0.26
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
ib-insync==0.9.86
ibapi==9.81.1.post1
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.1.3
mplfinance==0.12.10b0
msgpack==1.0.3
multidict==6.6.3
multitasking==0.0.11
narwhals==1.46.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndsplines==0.2.0
nest-asyncio==1.6.0
notebook==7.4.3
notebook_shim==0.2.4
numpy>=1.26
oauthlib==3.2.2
opencv-python==4.11.0.86
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandas_market_calendars==5.1.1
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.2
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.22.1
prompt_toolkit==3.0.47
propcache==0.3.2
protobuf==6.31.1
psutil==6.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyerfa==2.0.1.4
Pygments==2.18.0
pyluach==2.2.0
pyparsing==3.1.2
pyperclip==1.9.0
pytesseract==0.3.13
python-amazon-paapi==5.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
schedule==1.2.2
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
simupy==1.1.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sseclient-py==1.8.0
stack-data==0.6.3
streamlit==1.46.1
ta==0.11.0
telegram==0.0.1
tenacity==9.1.2
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
traitlets==5.14.3
tweepy==4.15.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.20
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
yarl==1.20.1
yfinance==0.2.65
